IndexingQueues: Simplify the iteration logic
Simplify IndexingQueue::processNextIteration(). It no longer requires
the queue to return a bool value indicating if the queue started
processing a file. Instead, it's the queue's responsibility to call
finishIteration() when they are done with processNextIteration().

This is much easier to understand. I would always get confused what the
boolean value was for.