FileIndexer: Improve the EventMonitor integration
Rewritten the EventMonitor to just monitor for events, and inform the
world about these events via signals. The IndexScheduler, now contains
the EventMonitor and reacts to those signals.

The Index Scheduler has 4 states -
* Low Disk Space
  -> All queues are suspended

* User Idle
  -> All queues are resumed

* On Battery
  -> Basic queue is resumed
  -> File Indexing queue is paused

* Normal
  -> Basic queue is resumed
  -> File Indexing queue is resume (should be configurable)

The IndexScheduler can also be externally suspended and resumed, which
will affect all the queues.

This commit temporarily removes all of the Event notifications.