IndexScheduler: Allow the user to configure the behaviour of the queues
One can now set a delay in between items of the queue, and one can
configure if one wants the file indexing queue to be paused during
normal execution.

In 4.9, the indexing queue used to have a large delay when indexing
normally. In 4.10, there are now 2 queues - basic and file. The file
indexing queue is, by default, suspended during normal execution. Added
an option to allow the user to change that.

In nepomukstrigirc -
[Indexing]
BasicIQDelay = 0
FileIQDelay = 0
NormalMode_FileIndexing = "suspend" | "resume"