DMM: Optimize removeTrailingGraphs query
The earlier query was complicated and forced virtuoso to count the
number of statements in each of the graphs provided, instead of just
checking for the existance of statements in those graphs. We were using
this counting query cause the existance query didn't seem to be working.

However, with virtuoso advancements, the query seems to now be working,
and we can even use the simple FILTER NOT EXISTS syntax, instead of the
ugly FILTER(!bif:exists(..)) syntax.