DMM: Auto-commit after each data manipulation statement
With the ResourceMerger rewrite we no longer push one statement at a
time. We now push them in one large statement. This has been shown to
cause transaction deadlocks in virtuoso. It is recommended to set
log_enable( 3 ) in order to auto commit after each data manipulation
command.

Ideally we should have proper transaction support and each command in
the data management model should be a separate transaction. However,
this will have to do until we figure out how to do that.

REVIEW: 107637