FileIndexer: Give plugins better control over indexing
Plugins could initially only supply a list of mimetypes which they
support, and they would then be called when a file of that mimetype was
found.

Now, they can implement a function shouldIndex which allows them to
control exactly when they should be called. Both the old and the new
mechanism are supported. Plugins need to report which mechanism they are
using.

Also implemented an ExtractorPluginMangager to handle getting the list
of plugins to call for a particular url and mimetype.