BasicIndexingQueue: Never reindex directories
The directory metadata never really changes. The only thing that changes
is the mtime of the directory. As an optimization we do not re-index a
directory after it has been indexed once.

This also helps us by not removing any nie:isPartOf relations that may
be present in the folder.