Fix title extraction for the popplerextraction
Extend popplerextractor with firstpage parsing

Often the pdf metadata is not available or wrong data is added
to the title field (pdf exporter names instead of title).

This patch adds the possibility to parse the first page for a possible
title. A possibel title is determined by the connected text with the
biggest font that was more than one character.

REVIEW: 107870