FileIndexer: Do not run the IndexCleaner and IndexingQueues in parallel
The IndexCleaner and IndexingQueues now no longer run in parallel, the
queues only run once the cleaner has finished. This is important cause
we now have a two stage indexing, so the fileIQ could continue to index
files whose contents will soon be cleared, or once they are cleared, or
worse while a file is being cleared (We do not have transaction support,
so this could end very badly)

This also solves the bug where the indexer wouldn't immediately abandon
indexing of a particular directory when the config file changed.

BUG: 306341