FileIndexerConfigTests: Make sure the tests always pass
Depending the user's KDE configuration, KTempDir may create a directory
in a hidden folder. If that is the case, then all the tests will fail,
cause they explicitly check for hidden folders. In that case we rather
create a temporary directory in /tmp instead.

It's not perfect, but it works.