Add queue and job classes for the webminer
This commit adds the queue and job classes to the indexscheduler.
Whenever a file file is indexed and as soon as the fileinder did its
job (kext:indexinglevel == 2) the webimer in called on the resource.

Everytime the webminer was called on a resource the indexinglevel 3 will
be set and the resource will not be requested again.

This initial commit needs the latest nepomuk-webminer from extragear to
be installed, as this executable will be called in the webminerjob via
QProcess.

The idea behind this approach is to replace the current Nepomuk2:Service
with the structure used by the indexer /suspend/resume on low battery
and such) idially the webminer should only be called on supported
mimetypes instead of every resource on the system.