FileIndexer: Proper user status strings
The old indexer used to actually follow a directory based indexing
scheme, where at any time you would know which directory it was
indexing. This new one doesn't exactly follow that scheme, plus with the
2 queues, it makes reporting that information harder.

Therefore, I've changed the way the userStatusStrings are reported. This
new method provides less information, but at least it is accurate.