FileWatch: Do not always add inotify watches for removable media
Not everyone wants watches to be added to the removable media. They are
okay with loosing tags/ratings and in removable media. Also, certain
users do not watch watches ever added for Network Shares - It results in
a large amount of network load which is not required.

The following options can disable watches in nepomukstrigirc -

[RemovableMedia]
add watches=true
add watches network share=false

By default watches are never added to network shares, and they are
always added for removable media.

BUG: 308474
REVIEW: 108047