kinotify: Clean up orphaned EventMoveFrom cookies
When deleting a file via KDE it by default is moved into the the trash.
Since we do not install watches for the trash (a hidden directory), we
only receive the EventMoveFrom event, and do not receive the EventMoveTo
event. This causes the cookie hash to increase in size and never
decrease.

There if one moves a large number of files into the trash the filewatch
memory usage will increase and never decrease.

We now set a timer of 1 second after every EventMoveFrom event, and then
clear the cookie hash. 1 second is more than enough time for the
EventMoveTo event to occur.