QueryService: Use the ResultIterator instead of duplicating code
This had been initially done as a junior job but it had to be reverted
cause the ResultIterator used the ResourceManager::mainModel(). It can
now be applied cause we overrite the mainModel with our own model.

By default the mainModel is a localSocketModel - We do not need to
communicate over a socket cause we have direct access to the main
virtuoso model. We are the storage service!