QueryService: Decouple the SearchRunnables and the Folder class
The SearchRunnable classes used to contain a reference to the Folder and
call certain functions on it. Since the SearchRunnables were in another
thread they required mutexes and still often resulted in deadlocks and
crashes.

The new code uses signals and slots to communicate between the Folder
and the SearchRunnable classes. This is a lot more elegant and easier to
understand.

REVIEW: 108271