SimpleIndexing: Do not save the unix file ownner, group and permissions
They were for an optimizations in the nepomuksearch query slave where
these properties were fetched so that we did not have to stat the file.
However, now we no longer pre-fetch them cause we always need to stat
the file in order to check if it exists in the file system.