FFMPEG Extractor: Copy the video data to the main resource
The Video extractor saves most of the important data in sub-resources
which are connected to the main file resource via nie:hasPart. This
includes important stuff like video width, height, aspect ratio and
frame rate.

It would be better to copy this information to the main file resource as
well. Specially since displaying it is a lot easier that way.