FileIndexerConfig: Do not use the RemovableMediaCache
It doesn't make any sense to load the list of include/exclude folders
differently based on which devices are mounted. Specially cause this
list is used by the index cleaner to remove all the files which do not
satisfy the include critera.

So, when the device is umounted, the cleaner effectively removes all
traces of it.

Instead, we ALWAYS populate the cache with entries of the removable
devies as well as the normal include/exclude directories.