Remove the RemovableMediaModel
The RemovableMediaModel was extremely slow and resulted in slowing down
*all* the queries and results by over 20%. Additionally, due to various
optimizations in StoreResources, it used to save the indexed data as
normal file resources. There have also been a couple of cases of us not
discovering mounted media properly, which resulted in Nepomuk not saving
them with proper filex urls.

For 4.10, instead of this half-half approach, we're removing this
RemovableMediaModel completely. We will store all file urls with their
complete path instead. It's a lot faster and easier this way.