This patch calls slotScheduleIndexing whenever the basic indexing queue is finished.
Usually the fileindexer only starts when the KIdleTime reached 2min while the basic IQ starts right away when new files are found.
There does exist a config switch "NormalMode_FileIndexing=resume" but this only gets checked when the eventmonitor has some changes (mostly the idletime)

With this change if the user specifies the config switch, the fileindexer will index any new files right away.
If "NormalMode_FileIndexing" is not set or set as "suspend" the current indexing behaviour will not change.

REVIEW: 108421