Add fileIndexingDone dbus signal
In order to start the webminer right after the fileindexer did his job
this signal is needed.

Usually the service is started via the eventmonitor after the usual 2min
idletime. If there are no files with kext_indexinglevel 2 the scheduler
is stopped till the next 2min idle time kicks in and the scheduler is
executed again.

Now with this signal the webminer scheduler can look for new stuff once
new files are indexed.

This is extremly useful if indexing on Normal mode is enabled in the
nepomukstrigirc.

With this signal right after downloading a pdf/video the fileindexer
extracts its part and the webminer can be started right away, thus the
user can see immediately what nepomuk can do for them.

REVIEW: 108469