Query Service Folder: Avoid multiple refreshes of queries
When the storage has changed, the query is re-run. If the storage is
continuously being changed then each time it is changed the query is
re-run (Unless the old query was still running).

From now on, we use a timer and wait for 2 seconds after the storage has
changed in order to re-run the query. This has the negative impact of
not sending updates until there are no changes for over 2 seconds.

This might not happen in the case of indexing. However, this behaviour
is preferred over re-running the queries a large number of times.
Specially since the data mostly never changes.