ResourceData: Set the nao:prefLabel for new tags on store
When a new tag is created only its nao:identifier is used. However, most
people using Nepomuk seem to rely on Resource::label() to get the tag's
label. There is no convenient way to access the identifier.

Therefore, we might as well set both the label and identifier on new tag
creation.