IndexCleaner: Reduce the number of cleaning queries
The cleaner always checks for legacy data as well. This is not always
necessary as not everyone has legacy data. We have 2 kinds of legacy
data -

* configuration dependent
* independent

The independent data just needs to be removed once, after which it never
needs to be queried again. We check for this kind of data and switch on
a boolean config value when we're done. That way we never need to do it
again.

The "Configuration Dependent" data depends on which files are folders
have been marked to be indexed. For this kind of data we check if the
strigi indexGraphFor relation is present. If it is, then we must run the
queries. Otherwise, we ignore it.