FileWatch: ignore unnecessary ClosedAfterWrite signals
Whenever a file opened under write mode is closed, the ClosedAfterWrite
signal is emitted, even if the file has not been modified. This isn't a
problem as such cause the file indexer checks the mtime of the file
agains't the value in the database before indexing it.

However, as an optimization we can ignore all the ClosedAfterWrite
events, where the modification time of the file is less than 1 minute
from the current time.

This reduces the dbus traffic, and avoids checking the database each
time.