ActiveFileQueue: Immediately timeout when adding a file the first time
When a file is added the first time we instants timeout that file. If it
is added again within m_emittedTimeout seconds of the timeout, it is
then moved to the actual queue where it waits a certain amount of time
before it timeouts. If it is again queues during this wait, the timer is
reset.

The behaviour has been changed cause multiple enqueues of the same file
are the exception. We should not slow down the real-time indexing just
cause we could receive multiple events for the same file. With this
approach, both cases are suitably covered.