DMM: Resolve all the nodes before sending them forward
The resolveNode/Urls functions used to return a QHash which contain the
original -> resolvedNode. If it could not be resolved, then resolvedNode
would be blank.

With this patch, all the resolveNode/Url functions just return a list of
resolved nodes. If the node could not be resolved, it is then created at
that moment. It is created in the first graph that can be found
containing that application. This isn't correct according to the old
behaviour, but we just do not care about separate graphs any more.

This patch breaks a LOT of tests which I'm not going to fix right now.
When the entire graph based change have been done, then the tests can be
fixed. The current unit tests depend too strongly on the old graph
behavior.