DMM Tests: Fix add/setProperty tests
addProperty is a little less rigid. It now always updates does the
following -

* Create the application resource and its graph
* Update the nao:lastModified of the resource

Earlier the first operation was not performed if there was an error,
and the nao:lastModified was only updated if the added/set value was
different than the one that already existed. While those 2 behaviours
are "correct", they are a pain to implement correctly. The code is a LOT
simpler this way.

It doesn't matter that much if the nao:lastModified is updated when a
property is set that already exists. The same is the case for the
creation of the app graph. Overall, it doesn't matter that it gets
created even if the job results in an error. It will eventually get
used.

Changed the tests to reflect these two changes, and to reflect the new
graph approach where a statement is stored in multiple graphs if it is
maintained by multiple applications.

Also split testAddProperty into 3 tests.