Change dbus interface to the FileIndexer for status messages
To have a unified dbus interface this patch adds the signal:
* statusMessage(int, string)
and the method:
* statusMessage
* currentStatus

To expose the current state of the fileindexer via dbus.
This change results in the same dbus signal/slot that are used by
akonadi_nepomuk and the FileWatcher.

The integers for the status are the fixed enum values of the Nepomuk2::IndexScheduler::Status
State_Normal = 0,
State_UserIdle = 1,
State_OnBattery = 2,
State_LowDiskSpace = 3,
State_Suspended = 4,
State_Cleaning = 5

Therefore the following method s are marked deprecated:
* userStatusString - bad naming, same as statusMessage
* simpleUserStatusString - not used/returns the same as userStatusString

REVIEW: 108811