DMM StoreResources: Port to this new graph structure
This results in the statements being stored multiple times in different
graphs. The graph depends on the app and if the data is discardable.

Also, the nao:lastModified is always updated even if nothing has
changed. It's too much effort to fetch it and compare it.

Updated all the tests to reflect this multi graph approach.