Refactor nepomuk-simpleresource-rygen.py
To reduce to time necessary for large SimpelResource generation for
any application that needs more than a few ontologies.

This patch removes many calls to soprano and uses internal dictionary
handling to generate the classes. Thus the class generation for the
Nepomuk-WebMiner is down from ~25min to less than a second

REVIEW: 108916