Check DateTime for negative years in the taglib extractor
While QDateTime is perfectly happy with B.C. years, Virtuoso throws an
error if asked to add one.

So if we encounter a year < 0, which is almost certainly a broken tag,
replace it with year 1.

This fixes indexing on some files I had with broken tags (the year was
set to 2^32-1)

REVIEW: 108977