FileWatcher: remove unnecessary call to "shouldFolderBeWatched"
The IgnoringKInotify::filterWatch method, before anything else, checks
if the path should be watched according to the configuration (by calling
FileIndexerConfig::self()->shouldFolderBeWatched( path )). If not, it
returns false. But if the path should not be indexed, according to
FileIndexerConfig, it returns the result of another call to
"shouldFolderBeWatched", passing the same path. But this should always
return true, or the method would have returned false before.
Thus, the second call is unnecessary.

REVIEW: 109010