DataManagement: Make SubResource removal optional
The "RemoveSubResources" flag had been deprecated, and sub resource
handling had been made non-option. This patch re-introduces this
feature. This is being done, cause detecting all the sub-resources of a
set of resource is very expensive. One might not want to do that. It's
okay to have some resources which have been orphaned.