New Tool - nepomukshow
It can used to dump the contents of any resource to stdout. It has
additional flags which can be used to control if inferencing should be
enabled or not.

This is being imported from kde:scratch/vhanda/nepomuk-show. I don't
care that much about the history that much, so I'm not importing it.