Introduce Nepomuk Service 2
Nepomuk Service version 2 unlike the old services are not launched by
the nepomukservicestub process. Instead they each have their own
process, which they install.

Nepomuk Service2 are implemented by deriving from the Service2 base
class. The derived class needs to provide its own main function as well.

This patch implements the following -

* Nepomuk::Service2 implementation
* Port the nepomukserver to support both Service and Service2
* Port all the services to Service2
--- Storage (nepomukstorage)
--- File Indexer (nepomukfileindexer)
--- File Watch (nepomukfilewatch)

These Service2 classes are much lighter as some of them can be
QCoreApplications instead of a KApplication with a ui (The FileIndexer
requires a UI for KIdleTime)

Additionally, this makes debugging a lot simpler as we will no longer
get reports that "nepomukservicestub" crash or is acting up. We will
always know which service is at fault.