Move AsyncClientApiTest to the Nepomuk Test Library
Its main purpose is to test if the dbus interface is working. It
however, even tests the internals of the model, which others tests do.
Hence that code has been removed.

This patch removes all of the FAKE_DMS code, as it is now no longer
used.

Also, rewritten the asynclientapitest to use proper names and properties
instead of the abstract ones defined in the data management tests.