New Tool: NepomukCmd
This NepomukCmd tool is very similar to the sopranocmd tool execept for
the fact that it currently only supports the query command. Also,
inference is disabled by default. It can be enabled with the --inference
parameter.

We can later extend this tool to provide basic data management functions
as well.