consider "old" files
when a file that has a very old modification time is copied in an indexed folder, checking the modification time isn't enough.
we have to check the modification time of the folder *and* of the file, then index the file if any of them is recent.
in this case it catches both the cases an already existing file has been modified or the case when an "old" file gets copied there (in that case is the directory modification time that catches it)
BUG:316158