Storage: Add a resetRepository public function
This function closes the repository, and manually deletes the
repository. It then restarts virtuoso, and reloads the ontologies.

Additionally, this patch also ports the BackupManager to use this
resetRepository function instead of deleting all the statements
manually. Deleting all the statements is slow and happens in one
transaction which results in the entire repository being loaded in
memory. Also, the indexes aren't cleared when clearing all the
statements.

Also added some public functions to -
* closePublicInterfaces - local socket + query service
* openPublicInterfaces

This allows us to not let others access the repository when we're
clearing.

Also made the ResourceManager react to the storage service getting
deinitialized.