kinotify: Optimize the filterWatch process
The FilterWatch is used to to not add watches to everything.
Specifically we do not add watches to names in the exclude list. Also
we only monitor for creation and modification events in the include
folders.

FileIndexerConfig::shouldFileBeIndexed and shouldFolderBeIndexed have a
lot of duplicate code internally. They both use the expensive RegExp
cache.

It's better if we only use the RegExp cache once.

REVIEW: 110317