Migration: Convert into a SystemTrayIcon
The big ugly MigrationDialog is not launched by default. From now on we
check if there is any valuable data to backup and accordingly either
delete the repository or backup the tags + rating and restore it. Most
users shouldn't even notice the migration as it finishes within 5
minutes.

For the advanced users that care, they can launch the nepomukmigrator
--advanced, and run the actual migration where all the graphs are
merged. Normal users should not need to do that. Plus, it's terribly
slow.

Hopefully, this will make the migration painless.