BackupRestorationJob: Start the % with 3%
Gives the user a sense that something is happening when restoring a
backup.

Normally, when the user resets a backup a large amount of time is spent
in removing the old database and loading all the ontologies, so they do
not receive any indication when that is happening. This way they at
least see a '3%'