Directly connect to virtuoso for reads instead of going through the Storage Service
The Nepomuk Storage service no longer starts a local socket. All
communications now happen directly via ODBC to virtuoso. This results in
a 6-8x performance improvement.

We need to find a proper way of communicating the port number to connect
to and not use this hack, but otherwise this patch results in massive
performance improvements.

Writes still happen through the dbus interface.