IndexScheduler: Slow down the BasicIndexer if we have too many files
When the basic indexer runs on full speed, it consumes a LOT of cpu.
Virtuoso easily reaches about 80% of one core. This might be slightly
annoying to the user, hence we slow it down a little bit once we hit 200
files, and again on 500, 750 and 1000.

This is useful in the case when someone upgrades to 4.11 and their
Nepomuk database is migarted. At that point all of their files will be
reindexed. We do not want to use too much cpu.