FileIndexer: Disable initial update by default
Do not scan for new files on startup by default. It consumes a lot of
CPU and disk load. The only reason this scan was there was to make sure
all the files are up to date in the event that Nepomuk had been swtiched
off.

Since Nepomuk should always be running, we can safely disable this
initial update.