FileIndexer: Do not specify which file is being indexed
One can still request the current file via currentFile. It however does
not make much sense to send it along with the status string.

The user does not need to know about each file that is getting indexed.
Additionally, since we index over 15 files per second, this just leads
to a LOT of dbus traffic.