FileIndexer: Do not emit new status string when each file is done.
Since we are not showing the name of the indexed file anymore, we only
need to update the status string when it is actually done indexing,
instead of every time the file being indexed changes.

Without this patch the status string bounces constantly between "Indexing Files"
and "Scanning for changes", (because of short pauses between files being
indexed) which is distracting and unneccessary noise.

REVIEW: 110878