New Tool - nepomuksearch
This is a debugging tool which I have been using to debug search queries
and see how well virtuoso works. I think it's a good idea to ship this
by default since it will help users figure out if it's a problem with
Nepomuk or the application they are using.

Also, this shows search excerpts which are super useful!