FileIndexer: Do not attempt to index files multiple times
If a file indexing job fails, set kext::indexingLevel to -1, so
that the next time the file queue fills, it is omitted.
This means we don't burn CPU trying to index things multiple times.

Almost all of the patch is actually moving updateIndexingLevel from
indexer/indexer.cpp to utils.cpp, and updating the build system to
match.

REVIEW: 110794