MainModel: Guess the virtuoso port number better
This shouldn't affect normal users, but for developers who work on
Nepomuk, virtuoso isn't always started on port 1113 because a previous
running virtuoso. In that case when a new virtuoso instance is started,
it is started on the next available port - 1114.

All the applications still try to connect to 1113 because the socket
file still exists, and that causes problems.

We now try to connect to the highest port number after 1113 which has a
socket file.