DMM resolveUrls: Do not create file urls with nostat
The noStat option is used when removing resources and properties. If the
file does not exist in this case (and therefore we are not stating for
it) we do not want it to be created, and then removed. That's just a
waste of resources.

This caused a huge slowdown when deleting 1000s of files as a resource
would be created for each file and then deleted.

BUG: 320446