FileWatch: Do not delete metadata for files in a folder
When we receive an event to delete the metadata of a folder, we do not
need to query for all the files in that folder and delete the metadata
for those files.

Inotify gives us a separate event for each file that has been deleted
before it gives the folder deletion event. So the query being used was
rather useless.

This seems to have been implemented in 1b7ea56fbf0 in 2010 to mimic what
we were doing for file moves. Strange.