ActiveFileQueue: Use a QHash instead of a QList for the emitted entires
By default everything gets inserted in the emitted entries list, and
then when it appears the second time, it's forwarded to the actual
queue.

Since we need to do a find operation on the emitted entries, it's much
better to use a hash, instead of a list.

Tested by creating 10000 files