Kinotify: Handle overflow events
There are cases when we are not able to handle all of the events fast
enough, and an overflow event is generated. In that case, we just need
to continue processing events.

Prior to this patch, the path would get constructed as '/' and that
would lead to an assert being hit.

This behaviour can be reproduced by running the nepomukfilewatch in
valgrind and creating 10000 files. The event occurs somewhere between
the 9200 and 9500th file on my system.