Remove the org.kde.nepomuk.DataManagement service
It has always been the same as the Storage service. There is no point in
having 2 different names for the same thing.

Moreover, the errror messages now specify that org.kde.NepomukStorage
could not be found which is more descriptive than if you say
org.kde.nepomuk.DataManagement could not be found.