ResourceWatcher: Add resources, properties and types
Because ResourceWatcher::start is now no longer blocking, it could
happen that someone calls start and then calls addResource. Since there
is no interface the addResource would not result in a dbus call.

Hence, when the interface is finally registed we add all the resources,
properties and types that we did not add earlier.