Fix all the unit tests
QTEST_KDEMAIN_WITH_COMPONENTNAME overwrites the XDG_CONFIG_HOME env
variable with a custom value (.kde-unit-test/xdg/local) which does not
correspond to the location where the config files are stored. Thefore we
overwrite this value.

The unit tests were failing because each of the tests was trying to
fetch the virtuoso port number from the nepomukserverrc config file
which they could not find. Therefore they assumed virtuoso is not
running and failed.