Query: Filter out all non Nepomuk resource results via a regex
It's slow, but it needs to be done. It's better to do it in the database
instead of the clients doing it. Ideally, we should use graph groups,
but that seems to cause virtuoso to crash. Maybe virtuoso 7 will be
better.

The one negative point of this patch is that nepomuk:/me is now never
returned in the results.