Detect filenames (with an eventual regular expression)
"test.txt" and "test.*" are replaced with a comparison against NFO:fileName.

Note: Due to the architecture of the parser, this commit allows arbitrary
properties to be matched against regular expressions, as "sent by Jim*" will
be parsed to "sent by" "nfo:fileName=Jim.*" and finally to
"nmo::messageSender=Jim.*".