Support type-dependent properties
"named %1" does not have the name meaning for a file (file name) or for a contact
(full name). This commit adds support for such type-dependent properties and moves
code around. Utils::fuseTerms is now simpler and only fuses terms, while the
tuning of terms (interval comparisons, etc) is done in QueryParser.