Tests: robustify the sandbox
Running the nepomuk tests always timed out for me. I found that this was
because the nepomukserver was failing to start. There are two problems
here:

1) if the sandbox fails to start, it should just fail, not
timeout.

2) nepomukserver should not fail to start.

This patch fixes 1) by adding return value checking to the sandbox
script; if kdeinit4 & friends fail to start the test fails instantly

2) turned out to be because the nepomukserver could not
connect to the (local) X server. To fix this, I changed the script to be
started through xvfb-run, which starts a virtual X server for it to
connect to.

Finally, start the nepomukserver we just built, rather than the
installed one.

REVIEW: 112134