Backup Statement Generator: statements without a datatype can be valid
In Nepomuk the nao:description is stored as an rdfs literal, which means
it doesn't have a datatype, but is supposed to have a language tag.
Since we don't store the language tag, it is just stored as a string.
This used to be considered invalid by the statement generator and
therefore the nao:description (dolphin comments) was never backed up
even though it was requested.

Also added unit test

BUG: 323260
REVIEW: 111947