Provide user-friendly patterns with completion proposals
Allow translators to use regular expressions to match terms (this greatly eases
the task for highly declined languages), but hide them to the user.

If the user enters "moDified", the completion proposal will not be "modified at|on",
but "moDified at". When words are separated by pipes, only the first one is shown
to the user.