Storage: Only allow one storeResources job at a time
This effectively only allows one item to be indexed at a time. So, when
emails and files are being indexed in parallel, only one item will be
indexed at a time, the others will be queued up.

This is important because with 4.11, everything has to be reindexed and
that seems to be causing a lot of virtuoso cpu usage as both stages of
file indexing and email indexing happen in parallel.

DIGEST: Do not allow items to be indexed in parallel - The cpu usage for
virtuoso is just too high.